Adding new option 'globOptions'
`globOptions: { <any options for minimatch> }`

You can pass in globOptions and they will be applied to all of the globs being passed in via `globs` or `ignore`. This is a nice way to override some of minimatch's weirder default options (such as `dot`) without having to create instances of minimatches when using walksync.