Disable scheduled CI, and update Node versions
After 60 days, GitHub seems to disable the workflow.
https://github.com/joliss/node-walk-sync/actions/workflows/ci.yml says "This
scheduled workflow is disabled because there hasn't been activity in this
repository for at least 60 days", and there are no runs listed at all, so
perhaps GitHub is turning off not just the scheduled runs, but the whole
workflow. Let's see if turning off the scheduling brings it back.

This also replaces volta with actions/setup-node, as volta was failing on macOS.